misc: small fix or general refactoring i did not bother commenting
