+++ sdcard/rootfs/root/pwnagotchi/scripts/pwnagotchi/bettercap.py
